+++ ratpack-core/src/main/java/ratpack/server/internal/RequestBodyReader.java
@@ -21,6 +21,6 @@ import ratpack.exec.Promise;
 
 public interface RequestBodyReader {
 
-  Promise<ByteBuf> read(int maxContentLength);
+  Promise<ByteBuf> read(long maxContentLength);
 
 }
